feature: move badge to repositoy component
