+++ src/pages/products/[product]/repositories/[repository]/components/Header/Header.tsx
@@ -12,21 +12,13 @@ import { ProductFormData } from '@services/product';
 import { toNumber } from 'lodash';
 import { useProductQuery } from '@pages/products/hooks/useProductQuery';
 import GaugeSlider from '../GaugeSlider';
-const oneStarBadge = '/images/svg/badges/1stars.svg'
-const twoStarBadge = '/images/svg/badges/2stars.svg'
-const threeStarBadge = '/images/svg/badges/3stars.svg'
-const fourStarBadge = '/images/svg/badges/4stars.svg'
-const fiveStarBadge = '/images/svg/badges/5stars.svg'
-const zeroStarBadge = '/images/svg/badges/0stars.svg'
 
 function Header() {
   const { currentProduct, setCurrentProduct } = useProductContext();
-  const { currentRepository, latestTSQMI } = useRepositoryContext();
+  const { currentRepository } = useRepositoryContext();
   const { updateProduct } = useProductQuery();
 
   const [openModal, setOpenModal] = useState<boolean>(false);
-  const [showBadge, setShowBadge] = useState<boolean>(false);
-  const [badgePath, setBadgePath] = useState<any>('');
 
   const initialValues = currentProduct && [currentProduct.gaugeRedLimit, currentProduct.gaugeYellowLimit];
   const [values, setValues] = useState(
@@ -40,43 +32,6 @@ function Header() {
       setValues(initialValues);
     }
   }, [initialValues])
-
-  useEffect(() => {
-    setStars();
-  }, [])
-
-  const setStars = () => {
-    let star: any;
-    const value = latestTSQMI.value;
-
-    switch (true) {
-      case value == 0:
-        star = zeroStarBadge;
-        break;
-      case 0 < value && value < 0.2:
-        star = oneStarBadge;
-        break;
-      case 0.2 <= value && value < 0.4:
-        star = twoStarBadge;
-        break;
-      case 0.4 <= value && value < 0.6:
-        star = threeStarBadge;
-        break;
-      case 0.6 <= value && value < 0.8:
-        star = fourStarBadge;
-        break;
-      case 0.8 <= value && value <= 1.0:
-        star = fiveStarBadge;
-        break;
-      default:
-        star = null;
-    }
-
-    if (star) {
-      setBadgePath(star)
-      setShowBadge(true);
-    }
-  }
 
   const option = {
     series: {
@@ -156,9 +111,6 @@ function Header() {
         <Typography variant="caption" color="gray">
           {currentRepository?.description}
         </Typography>
-        <div>
-          <img src={badgePath} alt="Exemplo SVG" style={{ width: '120px', height: '25px' }} />
-        </div>
       </Box>
       <Box>
         <IconButton onClick={handleOpenModal}>
